+++ src/xyz/oihana/schema/constants/traits/business/DocumentsTrait.php
@@ -6,12 +6,17 @@ use xyz\oihana\schema\constants\traits\business\documents\AdjustmentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\AppliedEcoFeeTrait;
 use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentLineTrait;
 use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentTrait;
+use xyz\oihana\schema\constants\traits\business\documents\CreditNoteTrait;
+use xyz\oihana\schema\constants\traits\business\documents\DeliveryNoteTrait;
 use xyz\oihana\schema\constants\traits\business\documents\DocumentTotalsTrait;
 use xyz\oihana\schema\constants\traits\business\documents\EcoFeeRuleTrait;
 use xyz\oihana\schema\constants\traits\business\documents\InvoiceTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentInstallmentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentScheduleTrait;
 use xyz\oihana\schema\constants\traits\business\documents\QuoteTrait;
+use xyz\oihana\schema\constants\traits\business\documents\ReceiptTrait;
+use xyz\oihana\schema\constants\traits\business\documents\StatementEntryTrait;
+use xyz\oihana\schema\constants\traits\business\documents\StatementTrait;
 use xyz\oihana\schema\constants\traits\business\documents\TaxDetailTrait;
 
 /**
@@ -27,11 +32,16 @@ trait DocumentsTrait
         AppliedEcoFeeTrait        ,
         BusinessDocumentLineTrait ,
         BusinessDocumentTrait     ,
+        CreditNoteTrait           ,
+        DeliveryNoteTrait         ,
         DocumentTotalsTrait       ,
         EcoFeeRuleTrait           ,
         InvoiceTrait              ,
         PaymentInstallmentTrait   ,
         PaymentScheduleTrait      ,
         QuoteTrait                ,
+        ReceiptTrait              ,
+        StatementEntryTrait       ,
+        StatementTrait            ,
         TaxDetailTrait            ;
 }