+++ src/xyz/oihana/schema/constants/traits/business/documents/CreditNoteTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\CreditNote} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait CreditNoteTrait
+{
+    const string REASON             = 'reason' ;
+    const string REFERENCES_INVOICE = 'referencesInvoice' ;
+}