+++ src/xyz/oihana/schema/constants/traits/business/documents/DeliveryNoteTrait.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\DeliveryNote} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait DeliveryNoteTrait
+{
+    const string ORDER_DELIVERY = 'orderDelivery' ;
+}