+++ src/xyz/oihana/schema/constants/traits/business/documents/StatementEntryTrait.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\StatementEntry} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait StatementEntryTrait
+{
+    const string AMOUNT   = 'amount' ;
+    const string BALANCE  = 'balance' ;
+    const string DATE     = 'date' ;
+    const string DOCUMENT = 'document' ;
+}